+++ src/image/ThumbnailDimensions.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace mediaelch {
+
+/// @brief Represents the dimensions of a thumbnail, e.g. for episodes.
+struct ThumbnailDimensions
+{
+    int width = 400;
+    int height = 300;
+};
+
+} // namespace mediaelch