+++ CHANGELOG.md
@@ -38,6 +38,7 @@ https://mediaelch.github.io/mediaelch-doc/faq.html#where-are-mediaelchs-settings
  - Add experimental command line interface (#720)
  - Add userrating field
  - Add Kodi v16/v17/v18 switcher (#719)
+ - Make episode thumbnail size configurable in `advancedsettings.xml` (#776)
 
 ### Internal Improvements